% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixModel.R
\name{fixModel}
\alias{fixModel}
\title{Fix ARIMA model parameters}
\usage{
fixModel(x, verbose = FALSE)
}
\arguments{
\item{x}{a persephone object}

\item{verbose}{if TRUE the changed parameters will be reported}
}
\description{
Fixes the ARIMA model parameters to prevent re-estimation
}
