% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixOutlier.R
\name{fixOutlier}
\alias{fixOutlier}
\title{Fix outliers in the model}
\usage{
fixOutlier(x, timespan = 12, verbose = FALSE)
}
\arguments{
\item{x}{a persephone object}

\item{timespan}{number of months from the end of the time series where outliers are not fixed}

\item{verbose}{if TRUE the changed parameters will be reported}
}
\description{
Fixes detected outliers to prevent re-detection
}
