% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Qr_List.R
\name{generate_Qr_List}
\alias{generate_Qr_List}
\title{Generates a list containing the figures for a EUROSTAT quality report}
\usage{
generate_Qr_List(x)
}
\arguments{
\item{x}{an object of class \link{persephone}.}
}
\description{
Generates a list containing the figures for a EUROSTAT quality report
}
\keyword{internal}
