% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hchart_single.R
\name{hc_add_forecast}
\alias{hc_add_forecast}
\title{Add forecast to highchart (internal helper)}
\usage{
hc_add_forecast(hc, x)
}
\arguments{
\item{hc}{A highchart object}

\item{x}{Output object from seasonal adjustment}
}
\value{
A highchart object with forecast added
}
\description{
Add forecast to highchart (internal helper)
}
\keyword{internal}
