% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perHts.R
\name{hierarchicalTimeSeries}
\alias{hierarchicalTimeSeries}
\title{R6 Class for hierarchical time series}
\description{
Combine mutliple objects of persephone objects into a new
hierachical persephone object. On the resulting time series the user can perform direct and
indirect seasonal adjustments.
}
\examples{
\dontrun{
objX13 <- perX13(AirPassengers, "rsa3")

# change parameters for all series (iterate = TRUE)
ht <- perHts(a = objX13, b = objX13, method = "x13", template = "rsa0",
            automdl.enabled = FALSE,
            arima.p = 1, arima.d = 2, arima.q = 0,
            arima.bp = 1, arima.bd = 1, arima.bq = 0,
            iterate = TRUE)

ht$run()
ht

# change parameters only for aggregate series (iterate = FALSE)
ht2 <- perHts(a = ht, b = objX13, method = "x13", template = "rsa0",
             transform.fun = "Log", transform.adjust = NA,
             easter.enabled = TRUE,
             outliers = c("AO1953.7", "LS1954.6"),
             iterate = FALSE)
ht2$run()
ht2
}
}
\section{Super classes}{
\code{\link[persephone3:persephone]{persephone3::persephone}} -> \code{\link[persephone3:multipleTimeSeries]{persephone3::multipleTimeSeries}} -> \code{hierarchicalTimeSeries}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{components}}{the sub series of the hierarchical time series}

\item{\code{weights}}{the weights used for aggregating components}

\item{\code{indirect}}{whether to use direct or indirect adjustment}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ts}}{the aggregated time series}

\item{\code{params}}{of all components and the aggregated series}

\item{\code{paramsDirect}}{params of the aggregated series}

\item{\code{adjustedIndirect}}{results from the indirect adjustment where
all components are adjusted and then aggregated}

\item{\code{adjusted}}{results from the seasonal adjustment}

\item{\code{forecasts}}{get forecasts}

\item{\code{forecastsIndirect}}{get forecasts according to indirect adjustments}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hierarchicalTimeSeries-new}{\code{hierarchicalTimeSeries$new()}}
\item \href{#method-hierarchicalTimeSeries-run}{\code{hierarchicalTimeSeries$run()}}
\item \href{#method-hierarchicalTimeSeries-print}{\code{hierarchicalTimeSeries$print()}}
\item \href{#method-hierarchicalTimeSeries-iterate}{\code{hierarchicalTimeSeries$iterate()}}
\item \href{#method-hierarchicalTimeSeries-getComponent}{\code{hierarchicalTimeSeries$getComponent()}}
\item \href{#method-hierarchicalTimeSeries-updateParams}{\code{hierarchicalTimeSeries$updateParams()}}
\item \href{#method-hierarchicalTimeSeries-generateQrTable}{\code{hierarchicalTimeSeries$generateQrTable()}}
\item \href{#method-hierarchicalTimeSeries-clone}{\code{hierarchicalTimeSeries$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="persephone3" data-topic="persephone" data-id="diagnosticsTable"><a href='../../persephone3/html/persephone.html#method-persephone-diagnosticsTable'><code>persephone3::persephone$diagnosticsTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="persephone3" data-topic="persephone" data-id="fixModel"><a href='../../persephone3/html/persephone.html#method-persephone-fixModel'><code>persephone3::persephone$fixModel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="persephone3" data-topic="persephone" data-id="fixOutlier"><a href='../../persephone3/html/persephone.html#method-persephone-fixOutlier'><code>persephone3::persephone$fixOutlier()</code></a></span></li>
<li><span class="pkg-link" data-pkg="persephone3" data-topic="persephone" data-id="hchart"><a href='../../persephone3/html/persephone.html#method-persephone-hchart'><code>persephone3::persephone$hchart()</code></a></span></li>
<li><span class="pkg-link" data-pkg="persephone3" data-topic="persephone" data-id="plot"><a href='../../persephone3/html/persephone.html#method-persephone-plot'><code>persephone3::persephone$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="persephone3" data-topic="persephone" data-id="plotResiduals"><a href='../../persephone3/html/persephone.html#method-persephone-plotResiduals'><code>persephone3::persephone$plotResiduals()</code></a></span></li>
<li><span class="pkg-link" data-pkg="persephone3" data-topic="persephone" data-id="plotSeasIrrCal"><a href='../../persephone3/html/persephone.html#method-persephone-plotSeasIrrCal'><code>persephone3::persephone$plotSeasIrrCal()</code></a></span></li>
<li><span class="pkg-link" data-pkg="persephone3" data-topic="persephone" data-id="plotSiRatios"><a href='../../persephone3/html/persephone.html#method-persephone-plotSiRatios'><code>persephone3::persephone$plotSiRatios()</code></a></span></li>
<li><span class="pkg-link" data-pkg="persephone3" data-topic="persephone" data-id="plotSpectrum"><a href='../../persephone3/html/persephone.html#method-persephone-plotSpectrum'><code>persephone3::persephone$plotSpectrum()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hierarchicalTimeSeries-new"></a>}}
\if{latex}{\out{\hypertarget{method-hierarchicalTimeSeries-new}{}}}
\subsection{Method \code{new()}}{
create a new hierarchical time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$new(
  ...,
  list = NULL,
  method = c("tramoseats", "x13"),
  template = NULL,
  context = NULL,
  userdefined = NULL,
  spec = NULL,
  weights = NULL,
  recursive = TRUE,
  iterate = TRUE,
  speclist = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{one or more objects which are either of class persephone or
can be coerced to persephone objects with asPersephone. If more than
one element is supplied, the underlying time series must have the same
time instances. All elements supplied in ... must be named.

In addition to the ts/mts/persephone objects,
\code{...} accepts arguments for customizing the model specifications as alternative input to speclist.
These are passed to the rjd3 specification functions internally.}

\item{\code{list}}{a list of persephone objects as alternative input to \code{...}.
This argument can also handle mts objects}

\item{\code{method}}{specifies the method to be used for the direct adjustment
of the aggregate series. tramoseats or x13}

\item{\code{template}}{the name of the predefined specification template, passed
as the \code{name} argument to \code{\link[rjd3x13:x13_spec]{rjd3x13::x13_spec()}} or \code{\link[rjd3tramoseats:tramoseats_spec]{rjd3tramoseats::tramoseats_spec()}}.
Must be one of:
\code{"rsa0"}, \code{"rsa1"}, \code{"rsa2c"}, \code{"rsa3"}, \code{"rsa4"}, \code{"rsa5c"} for x13
\code{"rsafull"}, \code{"rsa0"}, \code{"rsa1"}, \code{"rsa2"}, \code{"rsa3"}, \code{"rsa4"}, \code{"rsa5"} for tramoseats.
Defaults to \code{"rsa3"}.}

\item{\code{context}}{passed as the \code{context} argument of \code{\link[rjd3x13:x13]{rjd3x13::x13_fast()}} or
\code{\link[rjd3tramoseats:tramoseats]{rjd3tramoseats::tramoseats_fast()}}, a list of external regressors (calendar or other) to be used for estimation.}

\item{\code{userdefined}}{passed as the userdefined argument to x13_fast() or
tramoseats_fast()}

\item{\code{spec}}{a model specification object of class JD3_X13_SPEC or JD3_TRAMOSEATS_SPEC
returned by \code{\link[rjd3x13:x13_spec]{rjd3x13::x13_spec()}} or \code{\link[rjd3tramoseats:tramoseats_spec]{rjd3tramoseats::tramoseats_spec()}}.}

\item{\code{weights}}{either a vector if the same weight is used for all time
points or a list of ts objects or a mts object if the weight varies
for different time points. They must have the same length as the
number of components.}

\item{\code{recursive}}{logical, if TRUE, apply this setting to all subseries as well. If FALSE,
only apply to aggregate}

\item{\code{iterate}}{logical, if TRUE apply the given ... spec/params also to all components}

\item{\code{speclist}}{a list of additional arguments for customizing the model specifications,
passed to the rjd3 specification functions internally. As alternative input to \code{...}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hierarchicalTimeSeries-run"></a>}}
\if{latex}{\out{\hypertarget{method-hierarchicalTimeSeries-run}{}}}
\subsection{Method \code{run()}}{
run the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$run(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{if \code{FALSE} (the default), the results of the run will
be returned invisibly}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hierarchicalTimeSeries-print"></a>}}
\if{latex}{\out{\hypertarget{method-hierarchicalTimeSeries-print}{}}}
\subsection{Method \code{print()}}{
print a hierarchical timeseries to screen
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hierarchicalTimeSeries-iterate"></a>}}
\if{latex}{\out{\hypertarget{method-hierarchicalTimeSeries-iterate}{}}}
\subsection{Method \code{iterate()}}{
iterate over all components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$iterate(
  fun,
  asTable = FALSE,
  component = "",
  unnest = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{a function that takes a persephone object as a parameter}

\item{\code{asTable}}{if true, the return value of this method will be coerced
to a data.frame}

\item{\code{component}}{the id of the component}

\item{\code{unnest}}{if \code{asTable = FALSE}, converts the return value from a
nested list into a flat list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
this function is similar to \code{lapply()} in the sense that it
can be used to apply a function to several persephone objects
simultaneously
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hierarchicalTimeSeries-getComponent"></a>}}
\if{latex}{\out{\hypertarget{method-hierarchicalTimeSeries-getComponent}{}}}
\subsection{Method \code{getComponent()}}{
extract a component series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$getComponent(componentId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{componentId}}{the id of a component}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hierarchicalTimeSeries-updateParams"></a>}}
\if{latex}{\out{\hypertarget{method-hierarchicalTimeSeries-updateParams}{}}}
\subsection{Method \code{updateParams()}}{
change parameters of components or aggregate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$updateParams(
  component = "",
  iterate = FALSE,
  speclist = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{component}}{character scalar name of the component to change; if empty, operate on aggregate}

\item{\code{iterate}}{logical: if TRUE, apply parameter changes to all components recursively}

\item{\code{speclist}}{list with named arguments to be changed}

\item{\code{...}}{named arguments to be changed (spec options forwarded to updateFun)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hierarchicalTimeSeries-generateQrTable"></a>}}
\if{latex}{\out{\hypertarget{method-hierarchicalTimeSeries-generateQrTable}{}}}
\subsection{Method \code{generateQrTable()}}{
Generate a table for the eurostat quality report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$generateQrTable(component = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{component}}{(optional) a sub-component to create the report for}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hierarchicalTimeSeries-clone"></a>}}
\if{latex}{\out{\hypertarget{method-hierarchicalTimeSeries-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
