% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{package}
\name{persephone3-package}
\alias{persephone3}
\alias{persephone3-package}
\title{persephone3: Object-Oriented Wrappers for the Seasonal Adjustment Functionality of the JDemetra+ R Ecosystem rjdverse}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package is an object oriented wrapper around the R Packages rjd3x13 and rjd3tramoseats, which performs time series adjustments with the java library JDemetra+.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statistikat/persephone3}
  \item Report bugs at \url{https://github.com/statistikat/persephone3/issues}
}

}
\author{
\strong{Maintainer}: Angelika Meraner \email{angelika.meraner@statistik.gv.at}

Other contributors:
\itemize{
  \item Alexander Kowarik \email{alexander.kowarik@statistik.gv.at} [contributor]
  \item Markus Fröhlich \email{markus.foehlich@statistik.gv.at} [contributor]
}

}
\keyword{internal}
