% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perTramo.R
\name{perTramo}
\alias{perTramo}
\title{tramoseats class}
\usage{
perTramo(
  ts,
  template = c("rsafull", "rsa0", "rsa1", "rsa2", "rsa3", "rsa4", "rsa5"),
  context = NULL,
  userdefined = NULL,
  ...
)
}
\arguments{
\item{ts}{a time series}

\item{template}{passed as the \code{spec} argument to \code{\link[rjd3tramoseats:tramoseats_spec]{rjd3tramoseats::tramoseats_spec()}}}

\item{context}{passed as the \code{context} argument of \code{\link[rjd3tramoseats:tramoseats]{rjd3tramoseats::tramoseats_fast()}},
a list of external regressors (calendar or other) to be used for estimation}

\item{userdefined}{passed as the \code{userdefined} argument to \code{\link[rjd3tramoseats:tramoseats]{rjd3tramoseats::tramoseats_fast()}}}

\item{...}{passed to \code{\link[rjd3tramoseats:tramoseats_spec]{rjd3tramoseats::tramoseats_spec()}}}
}
\description{
an R6 class that represents a tramoseats model
}
\section{Inherits}{
 \link{persephone}
}

\examples{
data(AirPassengers, package = "datasets")

obj <- perTramo(AirPassengers, "rsa3")
obj$run()
obj$ts
}
