% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeasIrrCalHc.R
\name{plotSeasIrrCalHc}
\alias{plotSeasIrrCalHc}
\title{Interactive Seasonal, Irregular, and Calendar Effects Plot}
\usage{
plotSeasIrrCalHc(x)
}
\arguments{
\item{x}{an object of class \link{persephone}}
}
\value{
An object of class \code{highchart}
}
\description{
The function generates an interactive time series plot of the seasonal
component, irregular component and calendar effects
for a \link{persephone} object together with one year forecasts.
}
