% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSiRatios.R
\name{plotSiRatios}
\alias{plotSiRatios}
\title{SI-Ratio Plot}
\usage{
plotSiRatios(
  x,
  plotEngine = c("hc", "ggplot"),
  main = NULL,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of class \link{persephone}.}

\item{plotEngine}{character. Either \code{"hc"} (default, interactive highcharter plots)
or \code{"ggplot"} (static ggplot2 plots).}

\item{main}{if \code{plotEngine="ggplot"}, plot title. Ignored for highcharter plots.}

\item{interactive}{If \code{plotEngine="ggplot"}, whether to wrap in plotly for
basic interactivity. Ignored for highcharter plots.}

\item{...}{other parameters passed to the underlying plot function. Not currently used.}
}
\value{
A highchart object (if \code{plotEngine="hc"}) or a ggplot/plotly object
(if \code{plotEngine="ggplot"}).
}
\description{
Create SI-ratio plots using either highcharter (interactive) or ggplot2 (static).
This is a unified interface for \link{plotSiRatiosHc} and \link{plotSiRatiosGgplot}.
}
\details{
For \code{x13Single} objects, displays SI-ratios, replaced SI-ratios, seasonal
factors, and their mean. For \code{tramoseatsSingle} objects, shows SI-ratios
and their mean.
}
\examples{
x <- perX13(AirPassengers)
x$run()
plotSiRatios(x, plotEngine = "hc")
plotSiRatios(x, plotEngine = "ggplot")
plotSiRatios(x, plotEngine = "ggplot", interactive = TRUE)

}
