% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSiRatiosGgplot.R
\name{plotSiRatiosGgplot}
\alias{plotSiRatiosGgplot}
\title{SI-Ratio Plot (ggplot2 version)}
\usage{
plotSiRatiosGgplot(x, main = NULL, interactive = FALSE, ...)
}
\arguments{
\item{x}{an object of class \link{persephone}.}

\item{main}{plot title}

\item{interactive}{If TRUE, wrap the \code{ggplot} object in \link[plotly:ggplotly]{plotly::ggplotly}
to add basic interactivity (zoom, pan, hover tooltips). If FALSE (default),
returns a static ggplot object suitable for export.}

\item{...}{other plotting parameters to affect the plot. Not currently used.}
}
\value{
Returns an object of class \code{ggplot} or \code{plotly}
}
\description{
Create a plot of SI-ratios (and, in case of x11, of seasonal factors) by
period using ggplot2, suitable for publications and reports
with easy export to PDF, PNG, and other formats. Supports optional plotly
conversion for basic interactivity.
}
\details{
Produces a \code{ggplot}/\code{plotly} object for objects of class \link{persephone}.
If the \code{persephone} object is of class \code{x13Single}, the function
generates a plot of the SI-ratios, the replaced SI-ratios, the seasonal
factors and the mean of the seasonal factors. For \code{tramoseatsSingle} objects,
the function plots SI-ratios and their mean.
}
\examples{
data(AirPassengers, package = "datasets")
# Generate a persephone object, in this case an x13Single object
obj <- perX13(AirPassengers, "rsa1")
obj$run()
# Plot the SI-ratios after run
plotSiRatiosGgplot(obj)

data(UKgas, package = "datasets")
# Generate a persephone object, in this case a tramoseatsSingle object
obj2 <- perTramo(UKgas, "rsa3")
obj2$run()
plotSiRatiosGgplot(obj2)

}
