% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSiRatiosHc.R
\name{plotSiRatiosHc}
\alias{plotSiRatiosHc}
\title{SI-Ratio Plot (Interactive)}
\usage{
plotSiRatiosHc(x)
}
\arguments{
\item{x}{An object of class \code{persephone}}
}
\value{
An object of class \code{highchart}
}
\description{
Interactive visualization of SI-ratios and seasonal factors by period using
highcharter. Supports zoom, pan, and hover tooltips for exploratory analysis.
For static plots suitable for publications, see \link{plotSiRatiosGgplot}.
}
\details{
For \code{x13Single} objects, displays SI-ratios, replaced SI-ratios, seasonal
factors, and their mean. For \code{tramoseatsSingle} objects, shows SI-ratios
and their mean.
}
\examples{
x <- perX13(AirPassengers)
x$run()
plotSiRatiosHc(x)
}
