% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_spec.R
\name{update_spec}
\alias{update_spec}
\title{Update rjd3-spec object for seasonal adjustment}
\usage{
update_spec(spec, method = c("x13", "tramoseats"), freq, spec_list)
}
\arguments{
\item{spec}{an rjd3-specification object.}

\item{method}{the seasonal adjustment method to be used, i.e. one of the characters "x13" oder "tramoseats".}

\item{freq}{the frequency of the time series, i.e. one of the integers 12 (months) or 4 (quarters).
Used for parsing outlier dates.}

\item{spec_list}{a list containing the additional arguments for customizing the model specifications.
Passed to the rjd3 specification functions internally. See \code{\link{perX13}} or \code{\link{perTramo}} for more details.}
}
\value{
An updated specification object (JD3_*_SPEC).
}
\description{
Internal specification dispatching for rjd3x13/rjd3tramoseats.
}
\keyword{internal}
