% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getDiagnostics}
\alias{getDiagnostics}
\title{Run multiple diagnostic extraction functions on a Persephone object}
\usage{
getDiagnostics(x, ..., funcs = c("getOutliers", "getArimaCoef"))
}
\arguments{
\item{x}{A Persephone object (e.g., \code{x13Single} or \code{tramoseatsSingle})}

\item{...}{Additional functions to run (passed as character strings)}

\item{funcs}{Character vector of function names to run (default: \code{c("getOutliers", "getArimaCoef")})}
}
\value{
Invisible list of results from each function
}
\description{
This convenience function allows running several diagnostic extraction
functions at once and printing their results. It is primarily used for
interactive exploration and debugging.
}
\note{
This is an internal function for interactive diagnostics.
}
\examples{
# obj <- perX13(AirPassengers, "rsa3")
# obj$run()
# getDiagnostics(obj)
# getDiagnostics(obj, funcs = c("getOutliers", "getArimaCoef", "gettsout"))
}
