% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{gettsout}
\alias{gettsout}
\title{Extract outliers and create time series for plotting}
\usage{
gettsout(x)
}
\arguments{
\item{x}{A Persephone object (e.g., \code{x13Single} or \code{tramoseatsSingle})}
}
\value{
A list with two elements:
\item{\verb{[[1]]}}{A \code{ts} object with columns for each outlier type (otlAO, otlLS, otlTC).
Values at outlier dates contain the original series values,
all other dates are NA.}
\item{\verb{[[2]]}}{A named list of Date objects, one per outlier type (AO, LS, TC),
containing the dates where outliers occurred. Used for dygraph
event annotations.}
}
\description{
This function processes outlier information from a Persephone object and
creates time series objects that can be used for plotting outliers on graphs.
It is primarily used internally by plot functions to display outlier events
(AO, LS, TC) on time series visualizations.
}
\note{
This is an internal function; use it only if you know what you're doing.
}
\examples{
# obj <- perX13(AirPassengers, "rsa3")
# obj$run()
# otl <- gettsout(obj)
# otl[[1]]  # time series with outlier values
# otl[[2]]  # dates for event annotations
}
\keyword{internal}
