% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persephone.R
\name{persephone}
\alias{persephone}
\title{persephone class}
\description{
Common interface for x13 and tramoseats
}
\section{Constructor}{

\itemize{
\item \code{new()}: Initializes an object. See the documentation for the derived
classes for details.
}
}

\section{Methods}{

\itemize{
\item \code{run(verbose = FALSE)}: Runs the model
\item \code{updateParams(...)}: Updates the current parameters
\item \code{plot(...)}: Interactive time series plot for a persephone object, see
\link{plot.persephoneSingle} or \link{plot.hierarchicalTimeSeries}
\item \code{plotSeasIrrCal(...)}: Interactive plot of the seasonal component,
irregular component and calendar
effects for a persephone object, see \link{plotSeasIrrCal}
\item \code{plotResiduals(...)}: Several interactive plots in connection with
residuals for a persephone
object, see \link{plotResiduals}
}
}

\section{Read-Only fields}{

\itemize{
\item \code{params}: A parameters object of class \code{SA_spec}. See \link[rjd3x13:x13_spec]{rjd3x13::x13_spec} and
\link[rjd3tramoseats:tramoseats_spec]{rjd3tramoseats::tramoseats_spec}.
\item \code{ts}: The time series.
\item \code{output}: The return value from the underlying functions \code{x13_fast} or
\code{tramoseats_fast}. This slot will
be empty (\code{NULL}) before \code{run()} is called for the first time.
}
}

\section{Parameters}{

\itemize{
\item \code{verbose}: should the output from the underlying rjd3x13/rjd3tramoseats
function be displayed after the
run? If \code{FALSE} (the default), the output will be returned invisibly.
}
}

\section{derived classes}{

\code{\link[=perX13]{perX13()}}, \code{\link[=perTramo]{perTramo()}}, \code{\link[=perHts]{perHts()}}
}

\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{params}}{A parameters object of class \code{SA_spec}. See \code{\link[rjd3x13:x13_spec]{rjd3x13::x13_spec()}}
and \code{\link[rjd3tramoseats:tramoseats_spec]{rjd3tramoseats::tramoseats_spec()}}.}

\item{\code{ts}}{the (unajusted) time series}

\item{\code{tsp}}{the \code{\link[=tsp]{tsp()}} of the underlying time series}

\item{\code{output}}{The return value from the underlying functions x13_fast() or
tramoseats_fast(). This slot will be empty (NULL) before run() is called for
the first time.}

\item{\code{context}}{passed to \code{\link[rjd3x13:x13]{rjd3x13::x13_fast()}} and \code{\link[rjd3tramoseats:tramoseats]{rjd3tramoseats::tramoseats_fast()}} when the
\verb{$run()} method is invoked}

\item{\code{adjusted}}{get the adjusted series}

\item{\code{adjustedDirect}}{see \link{perHts}}

\item{\code{spec}}{specifications passed to \code{\link[rjd3x13:x13]{rjd3x13::x13_fast()}} and \code{\link[rjd3tramoseats:tramoseats]{rjd3tramoseats::tramoseats_fast()}} when the
\verb{$run()} method is invoked}

\item{\code{forecasts}}{get forecasts from the model}

\item{\code{forecastsDirect}}{get direct forecasts from the model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-persephone-new}{\code{persephone$new()}}
\item \href{#method-persephone-run}{\code{persephone$run()}}
\item \href{#method-persephone-updateParams}{\code{persephone$updateParams()}}
\item \href{#method-persephone-plot}{\code{persephone$plot()}}
\item \href{#method-persephone-hchart}{\code{persephone$hchart()}}
\item \href{#method-persephone-plotResiduals}{\code{persephone$plotResiduals()}}
\item \href{#method-persephone-plotSeasIrrCal}{\code{persephone$plotSeasIrrCal()}}
\item \href{#method-persephone-plotSiRatios}{\code{persephone$plotSiRatios()}}
\item \href{#method-persephone-plotSpectrum}{\code{persephone$plotSpectrum()}}
\item \href{#method-persephone-print}{\code{persephone$print()}}
\item \href{#method-persephone-iterate}{\code{persephone$iterate()}}
\item \href{#method-persephone-generateQrTable}{\code{persephone$generateQrTable()}}
\item \href{#method-persephone-setOptions}{\code{persephone$setOptions()}}
\item \href{#method-persephone-clone}{\code{persephone$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-new"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-new}{}}}
\subsection{Method \code{new()}}{
create a new persephone object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
this function should not be used directly because this class is
abstract. Only initialize derived classes.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-run"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-run}{}}}
\subsection{Method \code{run()}}{
perform a run of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$run(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{whether to show outputs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-updateParams"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-updateParams}{}}}
\subsection{Method \code{updateParams()}}{
update parameters for the adjustment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$updateParams(template = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{x13_spec()} or \code{tramoseats_spec()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-plot"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-plot}{}}}
\subsection{Method \code{plot()}}{
visualize the results of an adjustment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{\link[=plot]{plot()}}
see \code{\link[=plot.persephoneSingle]{plot.persephoneSingle()}} or \code{\link[=plot.hierarchicalTimeSeries]{plot.hierarchicalTimeSeries()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-hchart"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-hchart}{}}}
\subsection{Method \code{hchart()}}{
visualize the results of an adjustment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$hchart(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{\link[=hchart]{hchart()}}
see \code{\link[=hchart.persephoneSingle]{hchart.persephoneSingle()}} or \code{\link[=hchart.hierarchicalTimeSeries]{hchart.hierarchicalTimeSeries()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-plotResiduals"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-plotResiduals}{}}}
\subsection{Method \code{plotResiduals()}}{
visualize residuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$plotResiduals(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{\link[=plotResiduals]{plotResiduals()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-plotSeasIrrCal"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-plotSeasIrrCal}{}}}
\subsection{Method \code{plotSeasIrrCal()}}{
visualize the seasonal component, irregular
component and calendar effects for a persephone object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$plotSeasIrrCal(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{\link[=plotSeasIrrCal]{plotSeasIrrCal()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-plotSiRatios"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-plotSiRatios}{}}}
\subsection{Method \code{plotSiRatios()}}{
visualize SI-ratios (and, in case of x11, of seasonal factors) by
period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$plotSiRatios(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{\link[=plotSiRatios]{plotSiRatios()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-plotSpectrum"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-plotSpectrum}{}}}
\subsection{Method \code{plotSpectrum()}}{
visualize the periodogram or the autogregressive spectrum.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$plotSpectrum(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{\link[=plotSpectrum]{plotSpectrum()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-print"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-print}{}}}
\subsection{Method \code{print()}}{
display a persephone object in the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-iterate"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-iterate}{}}}
\subsection{Method \code{iterate()}}{
see \link{perHts}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$iterate(fun, asTable = FALSE, unnest = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{see \link{perHts}}

\item{\code{asTable}}{see \link{perHts}}

\item{\code{unnest}}{see \link{perHts}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-generateQrTable"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-generateQrTable}{}}}
\subsection{Method \code{generateQrTable()}}{
create a table for the eurostat quality report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$generateQrTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-setOptions"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-setOptions}{}}}
\subsection{Method \code{setOptions()}}{
update options for the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$setOptions(
  context = NA,
  userdefined = NA,
  spec = NA,
  recursive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{context}}{see \code{\link[rjd3x13:x13]{rjd3x13::x13_fast()}} and \code{\link[rjd3tramoseats:tramoseats]{rjd3tramoseats::tramoseats_fast()}}}

\item{\code{userdefined}}{see \code{\link[rjd3x13:x13]{rjd3x13::x13_fast()}} and \code{\link[rjd3tramoseats:tramoseats]{rjd3tramoseats::tramoseats_fast()}}}

\item{\code{spec}}{see \code{\link[rjd3x13:x13]{rjd3x13::x13_fast()}} and \code{\link[rjd3tramoseats:tramoseats]{rjd3tramoseats::tramoseats_fast()}}}

\item{\code{recursive}}{only applicable to hierarchical series. propagates
the updates to sub-series. see \link{perHts}
#' @description fix the arima model
#' @param verbose if TRUE the changed parameters will be reported
fixModel = function(verbose = FALSE) {
if(is.null(self$output)){
warning("not run yet.")
return(invisible(NULL))
}
p <- self$output$regarima$specification$arima$specification$arima.p
d <- self$output$regarima$specification$arima$specification$arima.d
q <- self$output$regarima$specification$arima$specification$arima.q
bp <- self$output$regarima$specification$arima$specification$arima.bp
bd <- self$output$regarima$specification$arima$specification$arima.bd
bq <- self$output$regarima$specification$arima$specification$arima.bq
if(self$output$regarima$specification$arima$specification$enabled){
if(verbose){
message("The model", paste("(",p,d,q,") (",bp,bd,bq,") is now fixed."))
}
self$updateParams(arima.p=p,arima.d=d,arima.q=q,
arima.bp=bp,arima.bd=bd,arima.bq=bq,
automdl.enabled = FALSE
)

}else if(verbose){
message("The model", paste("(",p,d,q,") (",bp,bd,bq,") was already fixed."))
}
},
#' @description create a new single object
#' @param timespan number of months from the end of the time series
#' where outliers are not fixed
#' @param verbose if TRUE the changed parameters will be reported
fixOutlier = function(timespan = 12, verbose = FALSE) {
xxx <<- self
if(is.null(self$output)){
warning("not run yet.")
return(invisible(NULL))
}
from <- time(self$ts)\link{length(self$ts)-timespan+1}
y <- floor(from)
m <- as.character(round((from-floor(from))*self$tsp\link{3}+1))
if(frequency(self$ts)==4){
m <- c("01","04","07","10")\link{as.numeric(m)}
}
if(nchar(m)==1){
m <- paste0("0",m)
}
from <- paste0(y,"-",m,"-01")
if(self$output$regarima$specification$outliers$enabled){
possibleOutliers <- row.names(self$output$regarima$regression.coefficients)
possibleOutliers <- possibleOutliers\link[=substring(possibleOutliers,1,2)\%in\%
                                           c("AO","LS","TC")]{substring(possibleOutliers,1,2)\%in\% c("AO","LS","TC")}
if(self$tsp\link{3}==12){
outliers <- lapply(possibleOutliers, function(x){
x2 <- strsplit(x = substring(x,5,nchar(x)-1), split ="-")[\link{1}]
x2\link{1} <- ifelse(nchar(x2\link{1})==1,paste0("0",x2\link{1}),x2\link{1})
data.frame(type=substr(x,1,2),date=paste0(x2\link{2},"-",x2\link{1},"-01"))
})
}else if(self$tsp\link{3}==4){
outliers <- lapply(possibleOutliers, function(x){
x2 <- strsplit(x = substring(x,5,nchar(x)-1), split ="-")[\link{1}]
x2\link{1} <- c(I="01",II="04",III="07",IV="10")[x2\link{1}]
data.frame(type=substr(x,1,2),date=paste0(x2\link{2},"-",x2\link{1},"-01"))
})
}
if(!is.na(self$output$regarima$specification$regression$userdef$outliers\link{1})){
if(self$tsp\link{3}==12){
outliers <- outliers[sapply(outliers,
userdefOut = self$output$regarima$specification$regression$userdef$outliers\link{,1:2},
function(x,userdefOut){
m <- merge(x, userdefOut, by = c("type","date"))
return(nrow(m)==0)
})]

\if{html}{\out{<div class="sourceCode">}}\preformatted{  \}else if(self$tsp[3]==4)\{
    outliers <- outliers[sapply(outliers,
                                userdefOut = self$output$regarima$specification$regression$userdef$outliers[,1:2],
                                function(x,userdefOut)\{
                                  m <- merge(x, userdefOut, by = c("type","date"))
                                  return(nrow(m)==0)
                                \})]
  \}
  oldType <- self$output$regarima$specification$regression$userdef$outliers$type
  oldDate <- self$output$regarima$specification$regression$userdef$outliers$date
\}else\{
  oldType <- NULL
  oldDate <- NULL
\}
if(length(outliers)>0)\{
  if(verbose)\{
    for(i in seq_along(outliers))\{
      message(outliers[[i]]$type," outlier saved at ",outliers[[i]]$date,".")
    \}
  \}
  df <- data.frame(type=c(oldType,
                          sapply(outliers, function(x)x$type)),
                   date=c(oldDate,
                          sapply(outliers, function(x)x$date)))
  df <- unique(df)
  if(class(self)[1]=="hierarchicalTimeSeries")\{
    private$updateParamsDirect(usrdef.outliersEnabled = TRUE,
                               usrdef.outliersType = df$type,
                               usrdef.outliersDate = df$date)
  \}else\{
    self$updateParams(usrdef.outliersEnabled = TRUE,
                      usrdef.outliersType = df$type,
                      usrdef.outliersDate = df$date)
  \}

\}else\{
  if(verbose)\{
    message("No automatic outliers found.")
  \}
\}
if(verbose)\{
  message("Updating parameter outlier.from to '",from,"'")
\}
if(class(self)[1]=="hierarchicalTimeSeries")\{
  private$updateParamsDirect(outlier.from = from)
\}else\{
  self$updateParams(outlier.from = from)
\}
}\if{html}{\out{</div>}}

}else if(verbose){
message("Automatic outliers not enabled.")
}
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-clone"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
